// src/core/nodes/types.d.ts
export interface CanonicalNode {
  type: string
  name: string
  server: string
  port: number
  udp?: boolean
  tls?: boolean
  [key: string]: unknown
}